+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Provided a new implementation of `_.unique`, removed argument `_.isSorted`
 * `_.isNil` now returns true if arg is nil, an empty string or empty table and false otherwise.
 * `_.size` now returns 0 for empty args instead of `nil`
+* `_.unique` no longer accepts `iter` argument to transform original array values.
 
 #### Renamed
 * Renamed alias `_.uId` to `_.uid`